+++ Services/API/IProductData.cs
@@ -5,5 +5,5 @@ public interface IProductData
     public int Id { get; }
     string Name { get; }
     string Description { get; set; }
-    float Price { get; set; }
+    int Price { get; set; }
 }
