+++ Services/Data/ProductData.cs
@@ -7,9 +7,9 @@ public class ProductData : IProductData
     public int Id { get; }
     public string Name { get; }
     public string Description { get; set; }
-    public float Price { get; set; }
+    public int Price { get; set; }
     
-    public ProductData(int productId, string name, string description, float price)
+    public ProductData(int productId, string name, string description, int price)
     {
         Id = productId;
         Name = name;
